\name{excelFormula}
\alias{excelFormula}
\alias{ExcelFormula-class}
\alias{coerce,call,ExcelFormula-method}
\title{Create an Excel formula}
\description{
  This is a constructor for an Excel formula object represented in R.
  We can insert that into Excel spreadsheet documents as the formual for
  one or more cells.
}
\usage{
excelFormula(expr)
}
\arguments{
  \item{expr}{the expression either as a string or a quoted call or expression}
}
\value{

}
\references{
}
\author{
Duncan Temple Lang
}
\seealso{

}
\examples{

}
\keyword{programming}
\keyword{interface}

